tests/wpt/web-platform-tests/html/semantics/embedded-content/media-elements/interfaces/TrackEvent/createEvent.html

<!doctype html>
<title>TrackEvent created with createEvent</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<div id=log></div>
<script>
test(function(){
    // https://www.w3.org/Bugs/Public/show_bug.cgi?id=17268
    assert_throws_dom('NOT_SUPPORTED_ERR', function() {
        var ev = document.createEvent('TrackEvent');
    });
    var ev = new TrackEvent('foo');
    assert_false('initTrackEvent' in ev, 'initTrackEvent');
});
</script>
